requirements.txt
asgiref==3.5.0
dj-database-url==1.0.0
Django==4.0.2
django-cors-headers==3.13.0
django-heroku==0.3.1
djangorestframework==3.14.0
djangorestframework-oauth==1.1.0
djangorestframework-simplejwt==5.2.2
gunicorn==20.1.0
numpy==1.22.2
psycopg2==2.9.3
psycopg2-binary==2.9.3
PyJWT==2.6.0
python-decouple==3.6
python-dotenv==0.21.0
pytz==2022.2.1
sqlparse==0.4.2
whitenoise==6.2.0
